LayoutTests/fast/js/kde/completion-expected.txt

PASS foo(), i is 2
PASS caught is true
PASS successfullyParsed is true
